move damage stuff to its own file
